place CredentialType into separate src file
