+++ identity-enabler/deviceId-mobile-app/src/lib/helpers.ts
@@ -193,13 +193,3 @@ export function isExpired(credential: { expirationDate?: string }): boolean {
 export function wait(milliseconds: number) {
     return new Promise(resolve => setTimeout(resolve, milliseconds));
 }
-
-/**
- * Supported credential types in Zebra IOTA SDK.
- */
-export enum CredentialType {
-    PERSONAL_INFO = "PersonalInformationCredential",
-    HEALTH_TEST = "HealthTestCredential",
-    BLOOD_TEST = "BloodTestCredential",
-    DEVICE_ID = "DeviceIDCredential"
-}
