+++ identity-enabler/deviceId-mobile-app/src/lib/store.ts
@@ -1,5 +1,6 @@
 import { writable } from "svelte/store";
-import { CredentialType, persistent } from "./helpers";
+import { persistent } from "./helpers";
+import { CredentialType } from "../models/types/CredentialType";
 
 const hasSetupAccountInitialState = () => false;
 /**
