+++ identity-enabler/deviceId-mobile-app/src/models/types/CredentialType.ts
@@ -1,0 +1,9 @@
+/**
+ * Supported credential types in Zebra IOTA SDK.
+ */
+export enum CredentialType {
+    PERSONAL_INFO = "PersonalInformationCredential",
+    HEALTH_TEST = "HealthTestCredential",
+    BLOOD_TEST = "BloodTestCredential",
+    DEVICE_ID = "DeviceIDCredential"
+}
