+++ identity-enabler/holder-mobile-app/src/lib/helpers.ts
@@ -235,13 +235,3 @@ export function generateRandomNumericString(charLength: number): string {
     }
     return str;
 }
-
-/**
- * Supported credential types in Zebra IOTA SDK.
- */
-export enum CredentialType {
-    PERSONAL_INFO = "PersonalInformationCredential",
-    HEALTH_TEST = "HealthTestCredential",
-    BLOOD_TEST = "BloodTestCredential",
-    DEVICE_ID = "DeviceIDCredential"
-}
