+++ identity-enabler/holder-mobile-app/src/services/identityService.ts
@@ -1,11 +1,11 @@
 import Keychain from "../lib/keychain";
 import { generateRandomNumericString, parse } from "../lib/helpers";
-import type { CredentialType } from "../lib/helpers";
 import { account } from "../lib/store";
 import type { Identity, IdentityConfig } from "../models/types/identity";
 import * as IotaIdentity from "@iota/identity-wasm/web";
 import { CREDENTIAL_EXPIRY_DAYS, IDENTITY_WASM_PATH } from "../config";
 import { get } from "svelte/store";
+import type { CredentialType } from "../models/types/CredentialType";
 
 const {
     Client,
