+++ identity-enabler/verifier-mobile-app/src/lib/helpers.ts
@@ -152,13 +152,3 @@ export function isExpired(credential: { expirationDate?: string }): boolean {
 export function wait(milliseconds: number) {
     return new Promise(resolve => setTimeout(resolve, milliseconds));
 }
-
-/**
- * Supported credential types in Zebra IOTA SDK.
- */
-export enum CredentialType {
-    PERSONAL_INFO = "PersonalInformationCredential",
-    HEALTH_TEST = "HealthTestCredential",
-    BLOOD_TEST = "BloodTestCredential",
-    DEVICE_ID = "DeviceIDCredential"
-}
